requirements.txt
azure-ai-agentserver-agentframework==1.0.0b7
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
azure-ai-agentserver-core==1.0.0b7